# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API para radioxenter",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/paquetes/{cod_paquete}/servicios": {
      "get": {
        "tags": [
          "Paquetes"
        ],
        "description": "Endpoint para obtener todos los servicios de un paquete",
        "parameters": [
          {
            "name": "cod_paquete",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": [
          "Paquetes"
        ],
        "description": "Endpoint para crear un servicio de un paquete.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cod_servicio": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "cod_paquete",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Paquetes"
        ],
        "description": "Endpoint para editar los servicios de un paquete.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EditarPaqueteServicio"
            }
          },
          {
            "name": "cod_paquete",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/servicios/{cod_servicio}": {
      "get": {
        "tags": [
          "Servicios"
        ],
        "description": "Endpoint para obtener un servicio",
        "parameters": [
          {
            "name": "cod_servicio",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/servicios": {
      "get": {
        "tags": [
          "Servicios"
        ],
        "description": "Endpoint para obtener servicios",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Servicios"
        ],
        "description": "Endpoint para crear un servicio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Servicio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Servicios"
        ],
        "description": "Endpoint para editar un servicio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Servicio"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/tipoPrefEntrega/{cod_tipo_pref_entrega}": {
      "get": {
        "tags": [
          "Tipo Preferencia Entrega"
        ],
        "description": "Endpoint para obtener un tipo de preferencia entrega",
        "parameters": [
          {
            "name": "cod_tipo_pref_entrega",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tipoPrefEntrega": {
      "get": {
        "tags": [
          "Tipo Preferencia Entrega"
        ],
        "description": "Endpoint para obtener tipos de preferencia entrega",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Tipo Preferencia Entrega"
        ],
        "description": "Endpoint para crear un tipo preferencia entrega.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoPreferenciaEntrega"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Tipo Preferencia Entrega"
        ],
        "description": "Endpoint para editar un tipo preferencia entrega.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoPreferenciaEntrega"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/transacciones/{cod_transaccion}": {
      "get": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para obtener una transacción",
        "parameters": [
          {
            "name": "cod_transaccion",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/transacciones": {
      "get": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para obtener transacciones",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para crear una transacción.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Transaccion"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para editar una transacción.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Transaccion"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/transaccionServicios": {
      "get": {
        "tags": [
          "TransaccionServicio - DEPRECATED"
        ],
        "description": "Endpoint para obtener transacciones servicio",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "TransaccionServicio - DEPRECATED"
        ],
        "description": "Endpoint para crear una transacción servicio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TransaccionServicio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "TransaccionServicio - DEPRECATED"
        ],
        "description": "Endpoint para editar una transacción servicio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TransaccionServicio"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EditarPaqueteServicio": {
      "type": "object",
      "properties": {
        "servicios_paquete": {
          "type": "array",
          "items": {
            "type": "number"
          }
        }
      },
      "xml": {
        "name": "EditarPaqueteServicio"
      }
    },
    "Servicio": {
      "type": "object",
      "properties": {
        "cod_servicio": {
          "type": "number",
          "example": 0
        },
        "nombre_servicio": {
          "type": "string",
          "example": ""
        },
        "descripcion_servicio": {
          "type": "string",
          "example": ""
        },
        "precio_servicio": {
          "type": "number",
          "example": 0
        },
        "iva_servicio": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Servicio"
      }
    },
...
    "TipoPreferenciaEntrega": {
      "type": "object",
      "properties": {
        "nombre_tipo_pref_entrega": {
          "type": "string",
          "example": ""
        }
      },
      "xml": {
        "name": "TipoPreferenciaEntrega"
      }
    },
    "Transaccion": {
      "type": "object",
      "properties": {
        "documento_usuario": {
          "type": "number",
          "example": 0
        },
        "valor_transaccion": {
          "type": "number",
          "example": 0
        },
        "fecha_transaccion": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "nombres_acudiente": {
          "type": "string",
          "example": ""
        },
        "apellidos_acudiente": {
          "type": "string",
          "example": ""
        },
        "documento_acudiente": {
          "type": "number",
          "example": 0
        },
        "parentesco_acudiente": {
          "type": "string",
          "example": ""
        },
        "cod_entidad_doctor": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Transaccion"
      }
    },
    "TransaccionServicio": {
      "type": "object",
      "properties": {
        "cod_transaccion": {
          "type": "number",
          "example": 0
        },
        "cod_servicio": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "TransaccionServicio"
      }
    },
...
  }
}